refactor: Separates index.js file into modules
The main index.js file has been separated into each individual modules that will help with the understandability of the code at hand
