+++ index.html
@@ -111,7 +111,6 @@
       integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
       crossorigin="anonymous"
     ></script>
-    <script src="src/index.js"></script>
-    <script src="src/unitary-tests.js"></script>
+    <script src="src/index.js" type="module"></script>
   </body>
 </html>
